[media] MTWN-92 make players launch in their own processes
This CL makes media players run in their own instances of the
media_service process. media_service can now run both as a singleton
service and as a transient isolate. When the singleton service gets
a request for a player, it launches a transient instance and
delegates the request to that instance.

When a player is created, the client provides the video renderer.
Video renderers run in the singleton media_service process, so that
aspect of a player won't be in the isolate. A future CL will address
this issue.

Change-Id: I3f01d65506022890c31605638ed9879503df98b5