[kernel][tests] Fix AllocChecker test that leaks gigabytes of memory
alloc_checker_oom() was testing the case where an allocation fails
(where ac.check() returns false), but it was leaking a lot of memory
in doing so.  Fix that by replacing it with a simpler test,
alloc_checker_new_fails().

Change alloc_checker_new() so that it checks that constructors get
run, which alloc_checker_oom() was testing before.

ZX-1692 #done

Change-Id: I3fb57d3bea96f5fb0a944819c72b68e35dfd4941