[audio][usb] Maintain the initial-volume hack.
HW gain in USB audio devices can be controled in a number of different
ways, but the primary way is by manipulating the gain control settings
(when present) in the "Feature Unit(s)" which can exist as the part of
the graph of Units and Terminals which make up the USB audio device.

Right now, we have been trying to blindly just set all of the gain
control knobs for all of the Feature Units we discover to 100% gain.
While this may not always be the proper thing to do (see the comment
added in this CL), it is important to get the address of these Units
correct when attempting to communicate with them.  Specifically, the
descriptors for these units should exist in (and be addressed using
the ID of) the top level AudioControl interface descriptor which
defines the topology of the device, and not in the AudioStreaming
interface descriptors which have more to do with the host interface to
the USB audio device, not the control of the internal device topology.

For now, remember the proper interface ID for each of the Feature
Units we discover and use that when implement the hack, instead of
simply using the last interface ID we parsed.  See the TODO left in
this CL for details about how we should be moving forward in order to
eliminate the hack entirely.

Change-Id: Ibe6b5409cd14c87c6aa23726ba33fad86424e117