[sysinfo] always null-terminate buffer
The name returned by gethostname() may be truncated if the supplied
buffer is not large enough to store the full name and when that happens
it is unspecified whether the returned buffer includes a terminating
null byte. This commit ensures that the buffer is always null-terminated
before it is passed to a formatted output library function (printf() and
related functions).

See STR32-C. Do not pass a non-null-terminated character sequence to a
library function that expects a string in the SEI CERT C Coding Standard
for further reading.

Change-Id: I473f75b86d53b595a456bed0440ac417f786d882