[debugger] Add asynchronous line editing.
Adds a new implementation of interactive line editing with history and tab
completions. Currently this uses a simple blocking version, but the underlying
verson now supports pushing data to it.

Linenoise does not support this mode which we need for the debugger because the
UI needs to multiplex between reading stdin and reading from the pipe to the
agent.

Change-Id: I4439e430c307d9f50d8bec65650d34c863d8c4aa