[paver] make all x86 targets pass all kern parts
Users keep getting stuck in a way where they're failing to pave a new kernel
partition and it's causing boot failures for them. The "default" path of
no-args on x86 passes all images and unused images are consumed and skipped
(which is very fast - kernel images are small).

The correct future fix for this will be a smarter protocol that enables the
client to request the files it actually needs.

Change-Id: Iec02802ad2e7dad996da02b4d45f5f1386f79b1f