[magma] Add RetryAllocator
On some midgard GPUs, the GPU virtual address for a buffer needs to be
equal to the CPU virtual address, but less than 2**33. Since other
memory can occasionally but rarely be mapped below that point, we need
to succeed in mapping the memory before returning success from the
allocator. RetyAllocator takes a callback which actually maps the memory
so it can do that check.

MA-415

Change-Id: Idc6b120cf96a8aaf29c99cbcd3a8068797b992d5