[media] move video MediaRenderer to VideoRendererImpl
Video renderers are implemented by three cooperating classes:
VideoRendererImpl, VideoFrameSource and VideoConverter.
VideoRendererImpl implemented VideoRenderer and VideoFrameSource
implemented MediaRenderer.

The media player wants to be in charge of creating the video renderer
now that media players run in their own isolated processes. Part of
making this work involves allowing the client to get access to the
VideoRenderer interface fairly late and with an InterfaceRequest, not
an InterfaceHandle. That is, we want this:

    MediaPlayer.TakeVideoRenderer(VideoRenderer& r);

not

    MediaPlayer.TakeVideoRenderer() => (VideoRenderer r);

because the latter is much harder for the client to handle. This is
much easier to achieve if VideoRendererImpl implements both
MediaRenderer and VideoRenderer, hence this change.

Change-Id: I0e801c4b58005af1806aede94b3d3f37902975c9