[ddk][block] change device_rebind() semantics
Previously device_rebind() would call device_remove() on all of
the children of the device it was called on, *without* invoking
their unbind() hook, in violation of the device model.

Change this so that it instead causes all children to be unbound
and, then, when all those children are finally removed and eventually
released, *then* cause the rebind attempt.

This preserves the device model and further ensures that new children
do not get bound until the existing children are fully destructed.

Since all block protocol devices are accessed from userspace via the
shared block driver, instead of having the shared block driver forward
the BLOCK_RR_PART ioctl() to its parent, it fields this ioctl() itself,
causing its children to become unbound and the rebound.

One interesting change is that it no longer changes name -- if you RR_PART
against /dev/class/block/123, that device will persist (since it *is*
the same device, it's just that you asked for it to remove its children
and try again because its partition table may have changed, etc).

Currently we synthesize a REMOVE and ADD directory watch event for a
device being rebound.  We should instead just generate a MODIFY event,
but that will require changes to various clients (block-watcher, tests, etc)
That depend on the old behavior.

Handling of BLOCK_RR_PART is removed from all block core drivers.

Change-Id: I9fd40ec11c216e3d2e0e88ef65ccd23e031eb245