[audio] MTWN-89 Unify competing constants for max gain
This set the maximum gain for the audio renderer to 24.0 dB,
the current maximum value for the audio mixer core.

Also updated with this CL is the existing media_client API,
which contains (ironically) a copy of this constant also,
but this duplication was intentional, as the purpose of the
media_client API is to eliminate the need for those very
early clients to need to adopt the full FIDL interfaces.

Change-Id: I5761583ea38f3e7d623fd6607ee1ec86f3174ca7