[pkgfs] support binding to a start handle
Note: this only supports a single handle. It would be preferred to receive
two handles, one for /pkgfs and one for /system, but due to restrictions in
the way that the fdio dispatcher and thinfs rpc abstractions are built it is
not possible to bind more than one channel. As such, /system is provided on
the /pkgfs handle at /pkgfs/system. fshost can present /pkgfs/system as
/system to the relevant parties.

Change-Id: I7ee0e8392793be728bfb26efcca02ff8393ccdd5