[ddk][usb-request] Remove usb_request_phys()
USB request buffers are always non-contiguous,
and io_buffer_phys() now asserts when called on non-contiguous buffers.
So we should now require using usb_request_phys_iter_init()
and usb_request_phys_iter_next() instead.

The dwc2 and dwc3 drivers do not yet support scatter/gather
and require all transactions to fit in a page,
but that is a separate problem to fix.

Change-Id: Ib25df2d5dcea9d2ae8cda4c80783f4b5a6d82c00