[debugger] Hook up debugger message loop.
Hooks up the "ps" command to get the process list as a proof-of-concept.

Asynchronously multiplexes between reading commands from stdin and data from
the socket connected to the debug agent. This asynchronous system will need
some significant revisions but this is a start.

Hooks up message reading and dispatch on the client side of the socket. The
Session object tracks pending requests and issues callbacks when the response
to the message comes.

Adds the beginnings of the client object model (ClientObject) and implements
one object (the System object).

Changes the output system to send data to be printed to OutputBuffer in one
chunk, which allows the command line to be hidden while output is being
(asynchronously) printed to the screen.

Fixes a race condition in ExceptionHandler::OnSocketReadable where not all of
the data will be read if data comes between requesting the available bytes and
reading it.

Change-Id: I1e4fbe3be0b2d15b2d3efd0a77af5d49f3a7f8c6