[dev][usb-video] VideoBuffer implementation.
VideoBuffer tracks the list of offsets that are available to write to,
(free_frames_) and those that are locked (locked_frames_).

There is also an in_progress_frame_ field for the frame the driver
is currently writing to.

Since we don't call ReleaseFrame yet, I have left it to a future CL.

ZX-1129 #comment

Change-Id: I672dce13e5673b0f453db40753f2d30a4e807a04