Add Scudo as an opt-in memory allocator.
When config "is_scudo" is enabled, the Scudo allocator will be used.

e.g.:

fx set x86 --args use_scudo=true
fx full-build

(If someone has a strong opinion that this should be a variant rather
than a config setting, let me know--this seemed a little more
lightweight/cleaner and more like what we did with SafeStack, and since
we'd hope that eventually, like with SafeStack, we'll have this as the
default, it made more sense to me to keep it here.)

Tested on QEMU and some hardware devices and stuff seems pretty stable.

Change-Id: I8dc6204b71de1270980978cef13818045e13e0db