[modular] Fix StartDaisy(), StartDaisyInShell().
The method were implemented by calling StartModule() and
StartModuleInShell() from the callback of a ResolveModuleCall
Operation, which violates asynchronous atomicity, because the
methods schedule new Operations on the main operation queue,
with no guarantee that no other operations are scheduled in
between.

We change this to implementing StartDaisy() and StartDaisyInShell()
using atomic operations. We extend the existing AddDaisyCall operation
to support StartDaisy() and StartDaisyInShell() too. This also
fixes some inconsistencies around module_source values.

Change-Id: Ibd3fc44f312bb4dd1bda52f0ce0871aa0d1780b7