[devhost] simplify child/instance management a bit
Since we simplified instance lifecycle a while back instances
and children are similar enough that we can use a single list
for both.

This also means that instances can receive unbind() ops now,
which may allow further simplification of some types of drivers
using them.  Existing drivers should be unaffected.

Also tidy some code on a failure case to device add to avoid
touching a device after the possibly last downref of it.

Change-Id: I913408cd51eca766b162884f245e22011a62203a