[devhost] make use-after-release of a device quite fatal
We were not actually free'ing device_t's when their refcount
hit zero.  As a step toward that (and to introduce some more
robust testing for use-after-release) we introduce a purgatory
list for deleted devices, invalidate all their pointers and
handles, and point their ops at a table of functions which
will print a suitable error message and crash.

In the future we'll start actually freeing devices once they've
been on the list after a while to avoid long term memory leaks.

Change-Id: If28343910feb793b81028a6dbc904c24c05997c1