[fidl][go] Add support for out-of-line objects
This change modifies both the fidl2 golang backend as well as the
bindings to support encoding out-of-line objects. This was recently
enabled by making both size and alignment available by the frontend.

This change also introduces a fidl2 file into the bindings which is used
to generate bindings for test types.

TO-324

Change-Id: I78d6a146dd8d95b35570526a65ad5e17c524a122