[appmgr] Adapt to non-overlapping namespaces
Now that namespaces cannot overlap, we need to explicitly whitelist the
directories that we give to shell programs. We also need to prevent
granting access to all of /dev/class because some sandbox policies map
in specific device classes.

There's still a potential conflict if a program maps all of /system/data
because vulkan and other sandbox features map in specific subdirectories
of /system/data. However, that issue will resolve itself when we remove
the system package.

Finally, this CL stops mounting root_info_experimental into / because /
is no longer a remote file system. We'll need to think more about how to
expose this information when this feature is more mature.

Change-Id: I52d0a9a4766dc564a05647d074c3c397ab4c3f43