[perftest] Add library for reporting raw test results data in JSON format
This is the first part of a performance test library for Zircon.

This initial part produces data in the JSON format used by
garnet/bin/zircon_benchmarks/test_runner.cc and by the trace utility
(as described in garnet/docs/tracing_usage_guide.md).  Eventually the
zircon_benchmarks code in Garnet will be changed to use this new
perftest library.

The next part will be a perf test runner that runs a test case
repeatedly.

ZX-1715

Change-Id: Ib9c14b6eebaa206ca5e07f34fca0a696eb495921