[build] Place finalize_manifests.py stripped results by source
When finalize_manifests.py does its own stripping (due to TO-471
and TO-475), it places the stripped files into the stripped/
subdirectory of the build directory.  It named the files by their
output name in the manifest, which now collides for common names
like bin/app in many separate package manifests.  Change to
instead using the basename of unstripped file, which doesn't in
fact conflict--though it's possible conflicts could crop up in
the future, so TO-842 is a better long-term solution.

TO-832 #done

Change-Id: I245bcc6f021bdf5c18420799ff68bf9f06311429