[aml][clk] Refactor AXG clock driver into AXG and GXL clock drivers
The Meson-AXG and Meson-GXL clock gating registers have similar
interfaces but a different set of registers controlling clock gating.

Refactor the common components out into a meson-lib clock library and have
the drivers for AXG and GXL link in this library with a seperate set of
registers for each driver.

Change-Id: I01dbfe77794c0a6543150ed9d145a344e31f4761