[build] Zircon ARM builds are now unified
We remove the concept of "zircon_platforms", replacing it with
"zircon_boards". "zircon_boards" contains a list of boards that the Zircon
build produces for the target architecture. Kernel disk images and bootdata
are built for all boards of the target architecture.

There is still a specialization for arm64 cases where `zircon.bin` is not a
bootdata, where we can not produce a unified `zedboot.bin` that contains both
kernel and ramdisk.

The gen argument "--platform" is removed as it is no longer required.

Change-Id: I7d498e3d95f0a1411956d661fb891e82121202cb