[build-zircon.sh] Prepare for reunified Zircon builds
This is an incremental step towards going back to a single unified
build per architecture.  With this, all the scripts, recipes, and GN
code can go to just using `//out/build-zircon/build-${target_cpu}`
for everything (and `build-${target_cpu}-ulib` for ASan libraries).

This builds some redundant stuff for now, but remains compatible
with both old and new regimes during the transition.

Change-Id: I2570052690913fe611540bdc47a79dd61a600fed