[docs] Split module elements into module and load
This change to the symbolization spec splits module contextual
elements into "module" and "mmap" contextual elements. Additionally
this change adds an extra parameter for the module reletive address of
a segment (p_vaddr in ELF). Modules now come in flavors of which the
only defined flavor is "elf". The "mmap" contextual element also now has
flavors of which the only one currently defined is "load"

Change-Id: I41a70cd7f1cce6505e591387fd4d97880d7cadbe