[wlan][broadcom] Change error type to zx_status_t
int, int32_t, etc. are replaced with zx_status_t and 0 is
replaced by ZX_OK wherever appropriate.

Timeout functions now return uint32_t, and the value is always
assigned to a variable named "time_left" for clarity.

Functions that used to overload their return types (e.g. error and
length, error and pointer) now have an additional out parameter
to receive the non-return type. In functions that return a pointer,
the pointer is set to NULL on error conditions.

zx_status_t values are always compared explicitly to == ZX_OK or
!= ZX_OK or == -EWHATEVER, never "if (err) {" or "if (!err) {"
and certainly not "if (err < 0) {"

Several different variable names are used to store status:
err, ret, rv, error, status. The names were not unified in this CL.

Linux error codes (ENOMEDIUM, EIO, etc.) are not replaced; that's
for the next CL.

Test: "fx build brcmfmac" succeeds.

NET-436
NET-517

Change-Id: I02e7143e06275d8ae74ce91f93fa6a15471a6b20