[cpp] Set defaults for prebuilt_shared_library type
prebuilt_shared_library is a template defined in
cpp/prebuilt_shared_library.gni that is intended to be a transparent
wrapper for a shared_library target along with some additional targets
for SDK packaging. One aspect that isn't transparent is the set of
default configs, which is applied when the target is instantiated and
can't be transparently forwarded through.

This makes it so the following works:

if (foo) {
    target_type = "prebuilt_shared_library"
} else {
    target_type = "shared_library"
}

target(target_type, "bar") {
    configs += [ ... ]
}

Change-Id: If65232c1d3f3c53d8f8bdfcff32df9c5c65d9215