[bt][hci] Redo CommandChannel SendCommand queueing
SendCommand is refactored to have a simpler argument list, and the
command flow is restructured to make the queues clearer and more usable
for asynchronous commands.

We will now send multiple commands at once to the controller, up to the
number of commands the controller reports we can send in CommandStatus
and CommandComplete messages (Bluetooth 5.0, Vol 2 Part E, Sections 4.4,
7.7.3 and 7.7.4).

Only one callback is now used, which receives all event packets resulting
from the command sent.

Commands sent to CommandChannel are not guaranteed to be sent in order,
SequentialCommandRunner or the callback should be used to guarantee
ordering if it is necessary.

Event Handlers cannot be "masked" by an command with the same complete
event anymore:
 - SendCommand will return 0 if there is already an event handler for
   the requested completion event
 - AddEventHandler will return 0 if an asynchonous command is running
   with the requested event code.

Command completion event matchers are removed (they were unused).

NET-174 #done

Change-Id: Idf3bd8353ff06ff0e0729de816b6b3928c853b9a