[fidl][go] Add support for the string type
This change adds support for the FIDL2 string type in the generator
backend as well as the bindings. It also adds support for a nullable
string type and string types with a maximum size. Thus, this change also
sets up the architecture for other types to be both nullable and have a
maximum size.

TO-324

Change-Id: I60b2c5dd6ea5dbe7ca76a7f171eb1dc6bc1fdea2