[usb][audio] Fix a bad assert.
Audio stream driver channels permit multiple concurrent users, but
only the "privileged" user is permitted to change the audio format.
If a non-privileged user attempts to change the format, the driver is
supposed to simply deny the request with ERR_ACCESS_DENIED, then hang
up the phone on the caller.

The code here attempts to do all of this, but it also had a sanity
check in place which was meant to ensure that the channel that the
request came in on was different from the "stream_channel_" member of
the driver class (the current privileged stream, if any).

Instead, this assert was attempting to assert that the user getting
booted was the *same* as the privileged stream; IOW the exact opposite
of what we wanted to assert.

TL;DR?  Flip the sense of this check so it is correct.

Change-Id: I5a7f3df97da004055a5d02784d33db3f2fc8c5c1