[fidl] Fix coding of vectors with boring payloads
Boring types don't make it into the coding tables. A uint32 struct
member is elided. So is an array of them. Things that aren't boring
(maybe transitively) have pointers or handles to check.

Vectors of boring things are semi-boring. We need the payload bounds,
but can skip looking at the contents.

Change-Id: I7f4d256d770406040baf17919bb5ab48d595fb5e