[fs][fdio] Simplify the 'Mmap' operation, convert to 'GetVmo'
This operation used to have some weird semantics: an input
length parameter, which was largely ignored, and an output
"offset" parameter. The reason for these arguments was for
servicing vmofiles, which are handled client-side anyway.

Rename the 'Mmap' operation to 'GetVmo' (since that's what
most filesystem servers are implementing anyway) and avoid
passing unused arguments.

Change-Id: I3fe9de194c3ae9d14a1f7dd790adba7fec1ab0e0