[audio][test] Fix a slightly off test.
Now that audio is reporting its minimum lead time requirement
properly, it is techincally possible to have a minimum lead time
requirement of 0.  This happens when your renderer is not connected to
any physical output, only the throttle output.

This situation is the norm in the test environment where the virtual
hardware does not expose any virtualized audio hardware.

Change-Id: I2fce4d5e6f59b1f71ccaf777b3c1bc35d4d8e246