[audio][mixer] Change default param, internal func
The AccumCosine function is used to generate sinusoidal audio
content for testing purposes. Previously, by default it added
into the existing contents of buffers instead of overwriting,
which has turned out to be the less-frequent usage.

This CL changes the default to the common usage, and updates
all applicable users.

Change-Id: Ifae2450b7fa570812cfe9b33c5bb78b6d157c864