[vim][display] Implement display ioctls
The get_fb ioctl can grab the real framebuffer, so an unnecessary copy
is avoided. This code is mostly temporary until the real GPU display API
is implemented.

This is a reland of 1229a100c04db0ec1d3fb3fe545252f9d4e808ad with a fix
for cache flushing through the ioctl.

Change-Id: I3965399ac00fc23d580f84e33cfd604d2a232276