[debugger] Work on command syntax.
This transitions from the LLDB-like category+verb syntax to a verb with
optional noun-modifier syntax described in go/zxdb. This new syntax should
be able to more easily scale to more threads and processes.

There are various improvements to the command system including many fewer
global lists of things to convert between enums and strings.

The actual debugger part of the debugger is unchanged, this affects the
text-based frontend only.

Change-Id: I1c77f8faa4e6dadd72dbaf737b51ac826692cbe3