[fs] Remove artificial MAX_TXN_MESSAGES constraint
Previously, the block midlayer limited the number of
permitted messages to an arbitrary constant size, since
it pre-allocated space for a fixed size number of incoming
messages.

However, with the revamped block protocol, no such limit exists:
clients are free to write requests until the block FIFO
is full.

Remove this constant, and allow filesystem transactions to
grow to an arbitrary number of fragmented requests.

ZX-1918 #done

Change-Id: I4922e93a2e25533d1a2a31dec17d25c02398f162