[rust] Partial fix for null build
Got rid of compatibilty unit test rules, as they
are no longer needed and were triggering
rebuilds.

Actually build lib targets, since those artifacts
are used in the artifacts that depend on them.

Stop trying to append the integration test
dependencies to the main dependencies file
as this would create a dependency loop between
integration test and the library it is testing.

The remaining item needed to fix the null build
is an update to cargo to have it generate gn/ninja
compatible dependency files. In particular, these
files must have an output-directory relative
path for the target of the dependency file.

Change-Id: I1a34a95dbedfe9eb42811bdc6eec544bfdc75da2