[makesdk] Start using create_layout.py in makesdk.go to add components
This teaches makesdk.go to start using the create_layout.py script to
add SDK components in the new form to the tarball, so consumers can
start migrating over to the new path. Initially, this adds the
'zircon_legacy' group which has async and async-default (and soon other
libraries from zircon.

Change-Id: I28d86ee2acd1b1ebd1e6f32b9d0a532d86271411