[bitmap] Add reverse scanning
This CL adds RScan and RFind to complement Scan and Find, respectively.
The functions find the last occurences of a non-matching bit or run of
matching bits, instead of the first.

The interface for Scan and Find has also slightly changed: while
returning "bitmax" made some sense on failure before, returning "bitoff"
doesn't for the reverse versions since it would be a valid offset.
Returning "bitoff-1" is similarly problematic, as bitoff may be 0.
Instead, Scan now returns a bool, and Find does not set "out" in their
respective failure cases.

ZX-1781 #comment Fixed

Change-Id: Ied08f8c74bca17e7b537d16075a2005e00d536fc