[package] Add mechanism for adding firmware blobs to /boot
This adds a mechanism for adding firmware blobs to the /boot filesystem
from the Fuchsia build as requested by the Zircon team in
https://fuchsia-review.googlesource.com/c/zircon/+/130378#message-e2aa13ab1d1c16cc8898516c2192629910a426af

To model this (without encouraging putting random stuff into /boot)
this introduces a new type of 'package' that can contain boot firmware
images and nothing else. This template produces a boot manifest that is
added to the set of boot entries considered by finalize_manifests.py.

Change-Id: I2e707d8f5e9085185b653f2a05dcc27e7ae4a849