[wlan][mlme] Use uint64_t for BSS' timestamp
This CL is a result of (1) not using chronos and instead zircon's time
implementation, and (2) be consistent with the `timestamp` type. As of
now we use chrono's timepoint for timestamp representation in BSS and
uint64_t to represent timestamp in frames. With this CL uint64_t is the
only type used for timestamp.

Change-Id: Ib1de41bd815e6bd5b0182f4cc7497d08b3cb7cff