[fidl][cpp] Assert !is_null_ in access operators
fidl::VectorPtr and fidl::StringPtr use get(), operator->, and operator*
to give access to the underlying std::vector and std::string
respectively. These operators should not be called on a null fidl type,
to prevent surprises where the underlying std type is modified through
the operator but the fidl type remains null.

TO-859
TO-860

Change-Id: I5c865f2b2ced40b5efb188145522e3868cd8461d