[debugger] Enhance client object model.
Makes the client object model virtual so we can mock it out for testing. No
testing currently does this, however.

Replaces the "watch all transition" callbacks with a more regular observer
model for the system, targets, and processes. This should scale better.

Removes object ID tracking from the client model as this is really a "view"
parameter that belongs in the console frontend. Added a ConsoleContext object
to track the mappings between IDs and client objects, and maintain the active
process and active thread for each process.

Wire up this console context to the commands so that all commands receive the
context specified on the command line, which is a combination of explicit
process/thread and the implicit active one.

Hooks up active process switching ("process 2") and thread switching ("thread
2"). Process listing is also supported, although there is currently no way to
make more than one. Thread listing is not yet supported.

Hooks up the arguments to the run command so that any process can be run, not
just "ps".

Enhance number parsing to make it more strict.

Change-Id: I81732ab851d12138fcf2fd4acbc258187e95c55e