[fidl] Test for FIDL binding compatibility.
Currently tests only that C++ is compatible with itself.
Once this is in place, the owners of the other language bindings can add
servers to the test by implementing a server and adding it to
kServerURLValues.

See README.md for an explanation of the basic structure of the test.

FIDL-21 # comment: Test for FIDL binding compatibility.

Change-Id: Ifd1e6c14b4628f161dc8643ebcd9e9636d3b4530