[audio][mixer] Mix accuracy tests
This CL adds fidelity-based tests for rechannelization and accumulation.
If we conceptualize the audio mixer as the following stages:

Normalize - Rechannelize - Interpolate - Scale - Sum - Denormalize

... then the Rechannelize and Sum nodes are the only remaining nodes
that have not been explicitly targeted with tests that are enabled
in the CQ mix. With this CL, the fidelity (dB level and noise floor)
of our stereo-to-mono and mono-to-stereo rechannelization and our
accumulation are measured and guarded.

As with the previous two CLs, a subsequent CL may include more
time-consuming, exhaustive combinatorial testing of these factors.

Change-Id: Id724a83039f21f2822a8bfd7f46500a6d557bb2a