[kernel][tests] Avoid a possibility of flaky failure in a preempt_disable test
Remove an unnecessary assertion.

100usec is often enough that timer_set() returns before the timer has
run, but it's not always enough, in which case the EXPECT_EQ() would
fail.

Rather than increasing the wait time that we pass to timer_set(), we
can just remove this assertion, which I had originally added just to
make sure that the "while (!timer_ran)" loop works properly.

ZX-1690
ZX-1820

Change-Id: I133e1996a519e8f74075043c73ddb012a254fdc8