[test][hypervisor] Fix interrupt tests
Use interrupt 32 as the test vector. Previously 0 was used, which is
counted as an NMI and injected into the guest before it has set up the
IDT.

Also, a slight change to interrupt injection code. Vectors 0-31 are
exceptions and so should by labeled as hardware exceptions when passed
to the guest. However, breakpoint and overflow exceptions should be
software exceptions. Previously all vectors up to the max Intel defined
vector were reported as hardware exceptions (and also considered to be NMIs),
here we lower that to only include vectors below the user defined base (32).

Change-Id: I067a7befac3a65b2aede692aefcfc79b74536de0