[gn] Consolidate references to Zircon build directories
This adds the //build/config/fuchsia/zircon.gni import to give
access to common variables and templates for interfacing with
the Zircon build.  Everything that's a simple use of a Zircon
tool can just use the `zircon_tool_action()` template.  Every
other reference to the Zircon build now uses the variables
defined as build arguments in zircon.gni.

Change-Id: I64d77aeb666e0751d20960952b798cc865b6bd8c