[package] Fix system packages triggering spurious rebuilds
The package template was expanding to a target that declared
boot_manifest as an output, but didn't always generate that file. Thus
ninja thought this build was was dirty due to a missing output.  This
fixes the template to always emit a system_manifest and boot_manifest
file, with one of them always being empty.

This also drops the --optional-manifest flag from finalize_manifest.py,
since all of the inputs will always exist.

Change-Id: Ic7c42a7c468674e3238db297374c3360a2f352b2