[debugger] Attach to running processes.
Implements an "attach" command and IPC message to attach to a running process.

The verbs_run file is renamed to verbs_process since it's where the "new"
command lives, and where other things like "kill" will go in the future.

A "new" command is also added to create a new process context. Example of
running and attaching two processes:

  [zxdb] new
  Process 2 Stopped <no name>
  [zxdb] run /boot/bin/ps
  Process 2 Running 6618 /boot/bin/ps
  [zxdb] process 1
  Process 1 Stopped <no name>
  [zxdb] attach 2446
  Process 1 Running 2446 <no name>

Change-Id: I9daecceef2756451982ffbd74a578b317c9c552e