[x86][hypervisor] Reinstate EPT invalidation
We used to invalidate EPT caches when launching a guest due to the use
of APIC addresses, which was removed when we moved to X2APIC. However,
we still need to invalidate the cache to cover the case where a guest is
destroyed immediately before a new guest with the same EPTP is started.
This was occurring randomly when running hypervisor-test in a loop and
causing triple faults.

PD-73 #done

Change-Id: Id250f52d7e3023daafcba78b0966667b5f9c78ad