[tpm] Fix zircon memory corruption via I2C serial bus
A physically promixate attacker which sits on the I2C serial bus between
the host processor and the TPM chip is able to manipulate certain
response packets in order to trigger memory corruption in the Zircon
kernel. Fix by ensuring proper computation of buffer bounds.

Fixes ZX-1768.

Thank you to Jeremy Boone of NCC Group for reporting this issue.

Change-Id: I8b412c6194ec971b7abf562016fef5dd7aa1a039