[gn] Rename gn/.gn to gn/dotfile.gn
The name of this file doesn't actually matter to GN, because it's
only used via the //.gn symlink.  However, if it's called .gn in a
subdirectory and GN runs with //build/gn (or a subdirectory of it)
as current directory, it will think //build/gn is the root of the
build (i.e. //).  It's also just bad form to have important source
files in the tree start with . when they don't need to, because
people tend to overlook the files.

Change-Id: Ifc56654c28cffe2fad0ede08d52807e9a2a15ec6