[amber] Allow retry logic to be kicked
On startup, if Amber has never been initialized it tries to reach
the update server. If the server is not found it does exponential
backoff when trying again. This change allow the FIDL server in
Amber to try running the initialization code immediately. This
change greatly improves the ergonomics of the development push
package workflow.

PKG-24 #done

Change-Id: I918dd8f74e875a9841dd2e960ad8239d541971c2