[bt][integration] Sleep while polling
The lifecycle test currently busy polls for /dev/bt-host devices
in a loop. The loop goes for 1000 iterations however this is
likely to cause flakes if device creation and binding takes too
long (which can often be the case on QEMU).

The test no longer relies on a high iteration count but instead
sleeps for 1 seconds in each iteration up, polling up to a maximum
of 30 seconds.

NET-590 #done

Change-Id: If4b1c7247e6852111a3bc5ede805f317d0b1ce5d