[debugger] Forward exceptions to client.
When the debug agent catches an exception, it now forwards basic
information about the exception to the client, and leaves the excepting
thread blocked. The client current just prints out the thread
information, this will be hooked to frame information in a followup.

The client currently selects the excepting thread and sets it as the
active thread, printing out the thread state. In the future, this will
be hooked to show the current stack frame for interrogation.

Fixed a bug sending new thread notifications. The process koid wasn't
being serialized. A test for this was added.

On attach, the agent now sends the thread list without being explicitly
asked.

Added a simple test app (not compiled by default) for manual testing.
Currently this is set up for testing exceptions.

Change-Id: I2386201363a9d21698000496057bf8725eef9182