[rust] compute GN dependencies using Cargo inputs
Cargo tracks an accurate list of all of the rust dependencies that are used
by our crates. If the Cargo data is wrong, the crates will not build.

The GN labels for rust artifacts do not repeat all of their labels, as it is
cumbersome to maintain two sets of dependency paths in two build systems.

It is strictly necessary to ensure that all crates that depend on FIDL
generated code have a GN dependency graph, such that the rust source code
that the FIDL compiler generates is invoked before any rust builds that
depend on it.

This change adds a script that computes a list of GN labels based on Cargo
data from the tree.

There is a strict convention implied by this script that the directory in
which a rust crate lives contains a rust_library target that is named
the same as the cargo crate name it builds.

TO-897 #done

Change-Id: I928b5061e6b2e31d95340136b85d23c9baddb060