[rust] all crates must be declared
A build change is coming that will produce correct dependencies for all rust
artifacts based on parsing Cargo metadata. It requires that every in-tree
rust target has a present patch declaration, and that every crate directory
contains a rust_library or rust_binary label with the same name as the
directory it lives in.

Change-Id: I54ca6a473a73fd72ea7b175c7dcbd735a38b0ab8