[crypto] Add AES128-CTR
This CL adds AES128-CTR, which is MUCH faster than our current
AES256-XTS implementation, but suffers from catastrophic nonce reuse and
MUST be replaced.

It's worth repeating: this is a TEMPORARY solution to unblock zxcrypt
until we have a fast XTS implementation, or can substitute it with
something else.

ZX-1811 #comment Prerequisite to switching zxcrypt

Change-Id: I63386bb2ece2348736b8fdd0e20556a84937258c